requirements.txt
Django==1.9.2
django-copyright==1.0.0
django-debug-toolbar==1.4
django-el-pagination==2.1.1
django-jenkins==0.18.1
markdown2==2.3.0
Pillow==3.1.1
psycopg2==2.6.1
Pygments==2.1.1
sqlparse==0.1.18
wheel==0.29.0
